+++ afat/static/afat/javascript/afat-logs.js
@@ -1,4 +1,4 @@
-/* global afatSettings, _dateRender, fetchGet, DataTable, _removeSearchFromColumnControl */
+/* global afatSettings, DataTable, _removeSearchFromColumnControl */
 
 $(document).ready(() => {
     'use strict';
@@ -6,40 +6,18 @@ $(document).ready(() => {
     /**
      * DataTable :: FAT link list
      */
-    fetchGet({url: afatSettings.url.logs})
-        .then((data) => {
-            const dt = new DataTable($('#afat-logs'), { // eslint-disable-line no-unused-vars
-                ...afatSettings.dataTables,
-                data: data,
-                columns: [
-                    {
-                        data: {
-                            display: (data) => _dateRender(data.log_time.time),
-                            sort: (data) => data.log_time.timestamp
-                        }
-                    },
-                    {data: 'log_event'},
-                    {data: 'user'},
-                    {
-                        data: {
-                            display: (data) => data.fatlink.html,
-                            sort: (data) => data.fatlink.hash
-                        }
-                    },
-                    {data: 'description'}
-                ],
-                columnDefs: [
-                    {
-                        targets: 0,
-                        columnControl: _removeSearchFromColumnControl()
-                    }
-                ],
-                order: [
-                    [0, 'desc']
-                ]
-            });
-        })
-        .catch((error) => {
-            console.error('Error fetching FAT logs:', error);
-        });
+    const dt = new DataTable($('#afat-logs'), { // eslint-disable-line no-unused-vars
+        ...afatSettings.dataTables,
+        serverSide: true,
+        ajax: afatSettings.url.logs,
+        columnDefs: [
+            {
+                target: 0,
+                columnControl: _removeSearchFromColumnControl()
+            }
+        ],
+        order: [
+            [0, 'desc']
+        ]
+    });
 });